feat: AuthenticationManager added
